+++ CHANGELOG.md
@@ -33,12 +33,14 @@ accidentally triggering the load of a previous DB version.**
 * #5497 Allow named time_bucket arguments in Cagg definition
 * #5499 Do not segfault on large histogram() parameters
 * #5500 Fix when no FROM clause in continuous aggregate definition
+* #5544 Fix refresh from beginning of Continuous Aggregate with variable time bucket
 
 **Thanks**
 * @nikolaps for reporting an issue with the COPY fetcher
 * @S-imo-n for reporting the issue on Background Worker Scheduler crash
 * @kovetskiy and @DZDomi for reporting peformance regression in Realtime Continuous Aggregates
 * @geezhu for reporting issue on segfault in historgram()
+* @H25E for reporting error refreshing from beginning of a Continuous Aggregate with variable time bucket
 
 ## 2.10.1 (2023-03-07)